break up WaitingListPage
