+++ src/components/StudentDatabaseToolbar/index.ts
@@ -1,3 +1,4 @@
 export * from "./ActionsMenu";
 export * from "./SelectStudentDialog";
+export * from "./StudentDatabaseToolbar";
 export * from "./StudentFilter";
