+++ src/components/StudentList/index.ts
@@ -3,7 +3,10 @@ export * from "./AddImageButton";
 export * from "./Image";
 export * from "./PlacementList";
 export * from "./ProgressBox";
+export * from "./StudentCard";
 export * from "./StudentCardHeader";
+export * from "./StudentCardImage";
 export * from "./StudentDatabaseActions";
 export * from "./StudentInfo";
+export * from "./StudentList";
 export * from "./WithdrawButton";
