+++ src/components/WaitingList/index.ts
@@ -1,6 +1,9 @@
 export * from "./JumpToEntryDialog";
+export * from "./WaitingList";
 export * from "./WaitingListActions";
+export * from "./WaitingListCard";
 export * from "./WaitingListCardHeader";
 export * from "./WaitingListCounter";
 export * from "./WaitingListEntryInfo";
 export * from "./WaitingListFilter";
+export * from "./WaitingListToolbar";
