+++ src/components/WaitingListForm/index.ts
@@ -4,4 +4,5 @@ export * from "./FormPlacementExam";
 export * from "./FormPlacementExamItem";
 export * from "./FormWaitingListVaccine";
 export * from "./WaitingListForm";
-export * from "./WaitingListDupPhoneDialog"
+export * from "./WaitingListDupPhoneDialog";
+export * from "./WaitingListFormDialog";
